# Conflicts:
#	composer.json
#	composer.lock
#	src/Modules/Maintenance/MaintenanceControl.php
#	src/Modules/Stats/StatsControl.php
#	src/Modules/Store/StoreModel.php

+++ src/Controller/BasketRestController.php
@@ -43,6 +43,7 @@ final class BasketRestController extends AbstractFOSRestController
 	private const KILOBYTES_PER_MEGABYTE = 1024;
 	private const MAX_PICTURE_SIZE_BYTES = 60 * self::KILOBYTES_PER_MEGABYTE * self::BYTES_PER_KILOBYTE;
 	private const SIZES = [800 => '', 450 => 'medium-', 200 => 'thumb-', 75 => '75x75-', 50 => '50x50-'];
+	private const MAX_BASKET_DISTANCE = 50;
 
 	public function __construct(BasketGateway $gateway, BasketService $service, ImageService $imageService, Session $session)
 	{
@@ -68,7 +69,9 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function listBasketsAction(ParamFetcher $paramFetcher): \Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
+		}
 
 		$baskets = [];
 		switch ($paramFetcher->get('type')) {
@@ -83,11 +86,41 @@ final class BasketRestController extends AbstractFOSRestController
 		return $this->handleView($this->view(['baskets' => $baskets], 200));
 	}
 
-	private function throwExceptionIfNotLoggedIn()
+	/**
+	 * Returns a list of baskets close to a given location. If the location is not valid the user's
+	 * home location is used. The distance is measured in kilometers.
+	 *
+	 * Returns 200 and a list of baskets, 400 if the distance is out of range, or 401 if not logged in.
+	 *
+	 * @Rest\Get("baskets/nearby")
+	 * @Rest\QueryParam(name="lat", nullable=true)
+	 * @Rest\QueryParam(name="lon", nullable=true)
+	 * @Rest\QueryParam(name="distance", nullable=false, requirements="\d+")
+	 *
+	 * @param ParamFetcher $paramFetcher
+	 *
+	 * @return \Symfony\Component\HttpFoundation\Response
+	 */
+	public function listNearbyBasketsAction(ParamFetcher $paramFetcher): \Symfony\Component\HttpFoundation\Response
 	{
 		if (!$this->session->may()) {
 			throw new HttpException(401, self::NOT_LOGGED_IN);
 		}
+
+		$location = $this->fetchLocationOrUserHome($paramFetcher);
+		$distance = $paramFetcher->get('distance');
+		if ($distance < 1 || $distance > self::MAX_BASKET_DISTANCE) {
+			throw new HttpException(400, 'distance must be positive and <= ' . self::MAX_BASKET_DISTANCE);
+		}
+
+		$baskets = $this->gateway->listNearbyBasketsByDistance($this->session->id(), $location, $distance);
+		$baskets = array_map(function ($b) {
+			$basket = $this->gateway->getBasket((int)$b[self::ID]);
+
+			return $this->normalizeBasket($basket);
+		}, $baskets);
+
+		return $this->handleView($this->view(['baskets' => $baskets], 200));
 	}
 
 	private function getCurrentUsersBaskets()
@@ -161,13 +194,18 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function getBasketAction(int $basketId): \Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
+		}
 
 		$basket = $this->gateway->getBasket($basketId);
+
 		if (!$basket || $basket[self::STATUS] == Status::DELETED_OTHER_REASON) {
 			throw new HttpException(404, 'Basket does not exist.');
 		} elseif ($basket[self::STATUS] == Status::DELETED_PICKED_UP) {
 			throw new HttpException(404, 'Basket was already picked up.');
+		} elseif ($basket['until_ts'] < time()) {
+			throw new HttpException(404, 'Basket is expired.');
 		}
 
 		$basket = $this->normalizeBasket($basket);
@@ -234,23 +272,38 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function addBasketAction(ParamFetcher $paramFetcher): \Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
-
-		$description = $this->getValidatedDescriptionOrThrowException($paramFetcher);
-
-		$lat = $paramFetcher->get(self::LAT);
-		$lon = $paramFetcher->get(self::LON);
-		if (!$this->isValidNumber($lat, -90.0, 90.0) || !$this->isValidNumber($lon, 0.0, 180.0)) {
-			// find user's location
-			$loc = $this->session->getLocation();
-			$lat = $loc[self::LAT];
-			$lon = $loc[self::LON];
-			if ($lat === 0 && $lon === 0) {
-				throw new HttpException(400, 'The user profile has no address.');
-			}
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
 		}
 
-		$basket = $this->createAndReturnBasketOrThrowException($paramFetcher, $description, $lat, $lon);
+		// prepare and check description
+		$description = trim(strip_tags($paramFetcher->get(self::DESCRIPTION)));
+		if (empty($description)) {
+			throw new HttpException(400, 'The description must not be empty.');
+		}
+
+		$location = $this->fetchLocationOrUserHome($paramFetcher);
+
+		$contactTypes = $paramFetcher->get(self::CONTACT_TYPES);
+		if ($contactTypes !== null && \is_array($contactTypes)) {
+			$contactTypes = array_map('intval', $contactTypes);
+		}
+
+		$basket = $this->service->addBasket(
+			$description,
+			'',
+			$contactTypes,
+			$paramFetcher->get(self::TEL),
+			$paramFetcher->get(self::MOBILE_NUMBER),
+			$paramFetcher->get('weight'),
+			$location['lat'],
+			$location['lon'],
+			$paramFetcher->get('lifetime')
+		);
+
+		if (!$basket) {
+			throw new HttpException(400, 'Unable to create the basket.');
+		}
 
 		// return the created basket
 		$basket = $this->normalizeBasket($basket);
@@ -267,43 +320,6 @@ final class BasketRestController extends AbstractFOSRestController
 			&& ($lowerBound <= $value) && ($upperBound >= $value);
 	}
 
-	private function getValidatedDescriptionOrThrowException(ParamFetcher $paramFetcher)
-	{
-		// prepare and check description
-		$description = trim(strip_tags($paramFetcher->get(self::DESCRIPTION)));
-		if (empty($description)) {
-			throw new HttpException(400, 'The description must not be empty.');
-		}
-
-		return $description;
-	}
-
-	private function createAndReturnBasketOrThrowException(ParamFetcher $paramFetcher, $description, $lat, $lon)
-	{
-		$contactTypes = $paramFetcher->get(self::CONTACT_TYPES);
-		if ($contactTypes !== null && \is_array($contactTypes)) {
-			$contactTypes = array_map('intval', $contactTypes);
-		}
-
-		$basket = $this->service->addBasket(
-			$description,
-			'',
-			$contactTypes,
-			$paramFetcher->get(self::TEL),
-			$paramFetcher->get(self::MOBILE_NUMBER),
-			$paramFetcher->get('weight'),
-			$lat,
-			$lon,
-			$paramFetcher->get('lifetime')
-		);
-
-		if (!$basket) {
-			throw new HttpException(400, 'Unable to create the basket.');
-		}
-
-		return $basket;
-	}
-
 	/**
 	 * Removes a basket of this user with the given ID. Returns 200 if a basket
 	 * of the user was found and deleted, 404 if no such basket was found, or
@@ -317,18 +333,57 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function removeBasketAction(int $basketId): ?\Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
-		$this->removeBasketOrThrowException($basketId);
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
+		}
+
+		$status = $this->gateway->removeBasket($basketId, $this->session->id());
+
+		if ($status === 0) {
+			throw new HttpException(404, 'Basket was not found or cannot be deleted.');
+		}
 
 		return $this->handleView($this->view([], 200));
 	}
 
-	private function removeBasketOrThrowException(int $basketId)
+	/**
+	 * Updates the description of an existing basket. The description must not be empty. If the location
+	 * is not given or invalid it falls back to the user's home. Returns the updated basket.
+	 *
+	 * @Rest\Put("baskets/{basketId}", requirements={"basketId" = "\d+"})
+	 * @Rest\RequestParam(name="description", nullable=false)
+	 * @Rest\RequestParam(name="lat", nullable=true)
+	 * @Rest\RequestParam(name="lon", nullable=true)
+	 *
+	 * @param int $basketId ID of an existing basket
+	 * @param ParamFetcher $paramFetcher
+	 *
+	 * @return \Symfony\Component\HttpFoundation\Response
+	 */
+	public function editBasketAction(int $basketId, ParamFetcher $paramFetcher): \Symfony\Component\HttpFoundation\Response
 	{
-		$status = $this->gateway->removeBasket($basketId, $this->session->id());
-		if ($status === 0) {
-			throw new HttpException(404, 'Basket was not found or cannot be deleted.');
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
 		}
+
+		$basket = $this->findEditableBasket($basketId);
+
+		// prepare and check description
+		$description = trim(strip_tags($paramFetcher->get(self::DESCRIPTION)));
+		if (empty($description)) {
+			throw new HttpException(400, 'The description must not be empty.');
+		}
+
+		$location = $this->fetchLocationOrUserHome($paramFetcher, [self::LAT => $basket[self::LAT], self::LON => $basket[self::LON]]);
+
+		//update basket
+		$this->gateway->editBasket($basketId, $description, $basket[self::PICTURE], $location[self::LAT],
+			$location[self::LON], $this->session->id());
+
+		$basket = $this->gateway->getBasket($basketId);
+		$data = $this->normalizeBasket($basket);
+
+		return $this->handleView($this->view(['basket' => $data], 200));
 	}
 
 	/**
@@ -343,11 +398,20 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function setPictureAction(int $basketId, Request $request): \Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
+		}
 
 		$basket = $this->findEditableBasket($basketId);
 
-		$data = $this->getValidatedDataOrThrowException($request);
+		$data = $request->getContent();
+		if ($data === '') {
+			throw new HttpException(400, 'The picture data must not be empty.');
+		}
+		if (strlen($data) > self::MAX_PICTURE_SIZE_BYTES) {
+			$maxPictureSizeMegabytes = self::MAX_PICTURE_SIZE_BYTES / (self::KILOBYTES_PER_MEGABYTE * self::BYTES_PER_KILOBYTE);
+			throw new HttpException(400, 'The picture data must not exceed ' . $maxPictureSizeMegabytes . ' MB.');
+		}
 
 		//save and resize image
 		$tmp = uniqid('tmp/', true);
@@ -366,23 +430,11 @@ final class BasketRestController extends AbstractFOSRestController
 
 		//update basket
 		$basket[self::PICTURE] = $picname;
-		$this->gateway->editBasket($basketId, $basket[self::DESCRIPTION], $picname, $this->session->id());
+		$this->gateway->editBasket($basketId, $basket[self::DESCRIPTION], $picname, $basket[self::LAT], $basket[self::LON], $this->session->id());
 
 		$data = $this->normalizeBasket($basket);
 
 		return $this->handleView($this->view(['basket' => $data], 200));
-	}
-
-	private function getValidatedDataOrThrowException(Request $request)
-	{
-		$data = $request->getContent();
-		if ($data === '') {
-			throw new HttpException(400, 'The picture data must not be empty.');
-		}
-		if (strlen($data) > self::MAX_PICTURE_SIZE_BYTES) {
-			$maxPictureSizeMegabytes = self::MAX_PICTURE_SIZE_BYTES / (self::KILOBYTES_PER_MEGABYTE * self::BYTES_PER_KILOBYTE);
-			throw new HttpException(400, 'The picture data must not exceed ' . $maxPictureSizeMegabytes . ' MB.');
-		}
 	}
 
 	/**
@@ -396,26 +448,21 @@ final class BasketRestController extends AbstractFOSRestController
 	 */
 	public function removePictureAction(int $basketId): \Symfony\Component\HttpFoundation\Response
 	{
-		$this->throwExceptionIfNotLoggedIn();
+		if (!$this->session->may()) {
+			throw new HttpException(401, self::NOT_LOGGED_IN);
+		}
 
 		//update basket
-		$basket = $this->removePictureFromBasketAndReturnBasket($basketId);
-
-		$basket = $this->normalizeBasket($basket);
-
-		return $this->handleView($this->view(['basket' => $basket], 200));
-	}
-
-	private function removePictureFromBasketAndReturnBasket(int $basketId)
-	{
 		$basket = $this->findEditableBasket($basketId);
 		if (isset($basket[self::PICTURE])) {
 			$this->imageService->removeResizedPictures('images/basket/', $basket[self::PICTURE], self::SIZES);
 			$basket[self::PICTURE] = null;
-			$this->gateway->editBasket($basketId, $basket[self::DESCRIPTION], null, $this->session->id());
+			$this->gateway->editBasket($basketId, $basket[self::DESCRIPTION], null, $basket[self::LAT], $basket[self::LON], $this->session->id());
 		}
 
-		return $basket;
+		$basket = $this->normalizeBasket($basket);
+
+		return $this->handleView($this->view(['basket' => $basket], 200));
 	}
 
 	/**
@@ -429,29 +476,49 @@ final class BasketRestController extends AbstractFOSRestController
 	private function findEditableBasket(int $basketId): array
 	{
 		$basket = $this->gateway->getBasket($basketId);
-		$this->validateBasketIsEditableOrThrowException($basket);
 
-		return $basket;
-	}
-
-	private function validateBasketIsEditableOrThrowException($basket)
-	{
-		$this->throwExceptionIfBasketDoesNotExist($basket);
-		$this->throwExceptionIfBasketIsOwnedByOtherUser($basket);
-	}
-
-	private function throwExceptionIfBasketDoesNotExist($basket)
-	{
 		if (!$basket || $basket[self::STATUS] === Status::DELETED_OTHER_REASON
 			|| $basket[self::STATUS] === Status::DELETED_PICKED_UP) {
 			throw new HttpException(404, 'Basket does not exist or was deleted.');
 		}
-	}
-
-	private function throwExceptionIfBasketIsOwnedByOtherUser($basket)
-	{
 		if ($basket['fs_id'] !== $this->session->id()) {
 			throw new HttpException(401, 'You are not the owner of the basket.');
 		}
+
+		return $basket;
+	}
+
+	/**
+	 * Returns a location from the param fetcher in the 'lat' and 'lon' fields. If none
+	 * is given, it returns the default location or the user's home address, if the default
+	 * location is null.
+	 *
+	 * @param ParamFetcher $paramFetcher
+	 * @param array $defaultLocation a fallback value or null
+	 *
+	 * @return array the location
+	 *
+	 * @throws \HttpException if no location and no default location were given and the user's
+	 * home address is not set
+	 */
+	private function fetchLocationOrUserHome(ParamFetcher $paramFetcher, array $defaultLocation = null): array
+	{
+		$lat = $paramFetcher->get(self::LAT);
+		$lon = $paramFetcher->get(self::LON);
+		if (!$this->isValidNumber($lat, -90.0, 90.0) || !$this->isValidNumber($lon, -180.0, 180.0)) {
+			if ($defaultLocation !== null) {
+				return $defaultLocation;
+			} else {
+				// find user's location
+				$loc = $this->session->getLocation();
+				$lat = $loc[self::LAT];
+				$lon = $loc[self::LON];
+				if ($lat === 0 && $lon === 0) {
+					throw new HttpException(400, 'The user profile has no address.');
+				}
+			}
+		}
+
+		return ['lat' => $lat, 'lon' => $lon];
 	}
 }
